+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - Add console colors in remove command
 - Add init command
 - Add tests for create ticket
+- Add function defaultKidneyWorkspacePath
 
 ### Changed